+++ h/schemas/forms/accounts/__init__.py
@@ -3,10 +3,12 @@
 from __future__ import unicode_literals
 
 from h.schemas.forms.accounts.edit_profile import EditProfileSchema
+from h.schemas.forms.accounts.forgot_password import ForgotPasswordSchema
 from h.schemas.forms.accounts.login import LoginSchema
 
 
 __all__ = (
     'EditProfileSchema',
+    'ForgotPasswordSchema',
     'LoginSchema',
 )
